+++ Phidgets22.indigoPlugin/Contents/Server Plugin/digitalinput.py
@@ -4,14 +4,19 @@ import datetime
 import indigo
 
 from Phidget22.Devices.DigitalInput import DigitalInput
+from Phidget22.Devices.Log import Log
+from Phidget22.LogLevel import LogLevel
 from phidget import PhidgetBase
 
 class DigitalInputPhidget(PhidgetBase):
-    def __init__(self, isAlarm, onStateIcon, offStateIcon, *args, **kwargs):
+    def __init__(self, isAlarm, onStateIcon, offStateIcon,
+                 logRawStateChanges=False, *args, **kwargs):
         super(DigitalInputPhidget, self).__init__(phidget=DigitalInput(), *args, **kwargs)
         self.isAlarm = isAlarm
         self.onStateIcon = onStateIcon
         self.offStateIcon = offStateIcon
+        self.logRawStateChanges = logRawStateChanges
+        self._rawStateChangeSequence = 0
 
     def addPhidgetHandlers(self):
         self.phidget.setOnErrorHandler(self.onErrorHandler)
@@ -35,6 +40,16 @@ class DigitalInputPhidget(PhidgetBase):
         self.indigoDevice.updateStateImageOnServer(stateImage)
 
     def onStateChangeHandler(self, ph, state):
+        if self.logRawStateChanges:
+            self._rawStateChangeSequence += 1
+            Log.log(
+                LogLevel.PHIDGET_LOG_INFO,
+                "DigitalInput raw callback: device='%s' id=%s serial=%s "
+                "hubPort=%s channel=%s sequence=%s state=%s" % (
+                    self.indigoDevice.name, self.indigoDevice.id,
+                    self.channelInfo.serialNumber, self.channelInfo.hubPort,
+                    self.channelInfo.channel, self._rawStateChangeSequence,
+                    bool(state)))
         self.updateIndigoStatus(state)
 
     def actionControlSensor(self, action):